game: fix clang warning + uncrustify + docs ref #55

+++ src/game/et-antiwarp.c
@@ -6,7 +6,13 @@
 
 #include "g_local.h"
 
-// fixes spectator bugs
+/**
+ * @brief G_DoAntiwarp
+ * @param[in] ent
+ * @return
+ *
+ * @note Fixes spectator bugs
+ */
 qboolean G_DoAntiwarp(gentity_t *ent)
 {
 	// only antiwarp if requested
@@ -35,6 +41,11 @@ qboolean G_DoAntiwarp(gentity_t *ent)
 	return qtrue;
 }
 
+/**
+ * @brief etpro_AddUsercmd
+ * @param[in] clientNum
+ * @param[in] cmd
+ */
 void etpro_AddUsercmd(int clientNum, usercmd_t *cmd)
 {
 	gentity_t *ent = g_entities + clientNum;
@@ -52,7 +63,14 @@ void etpro_AddUsercmd(int clientNum, usercmd_t *cmd)
 	}
 }
 
-// G_CmdScale is a hack :x
+/**
+ * @brief G_CmdScale
+ * @param[in] ent
+ * @param[in] cmd
+ * @return
+ *
+ * @note G_CmdScale is a hack :x
+ */
 static float G_CmdScale(gentity_t *ent, usercmd_t *cmd)
 {
 	float scale = abs(cmd->forwardmove);
@@ -119,6 +137,10 @@ static float G_CmdScale(gentity_t *ent, usercmd_t *cmd)
 	return scale;
 }
 
+/**
+ * @brief DoClientThinks
+ * @param[in,out] ent
+ */
 void DoClientThinks(gentity_t *ent)
 {
 	usercmd_t *cmd;
@@ -223,7 +245,7 @@ void DoClientThinks(gentity_t *ent)
 			// try to split it up in to smaller commands
 
 			delta     = ((float)LAG_MAX_DELTA - ent->client->cmddelta);
-			timeDelta = ceil(delta / speed); // prefer speedup
+			timeDelta = (int)(ceil((double)(delta / speed))); // prefer speedup
 			delta     = (float)timeDelta * speed;
 
 			if (timeDelta < 1)
